refactor edit collection form structure
