--- packages/web/src/components/create-playlist/CreatePlaylistButton.module.css
@@ -1,37 +1,0 @@
-.createPlaylistButton {
-  background: none;
-  border: none;
-  padding: 0;
-  cursor: pointer;
-
-  color: var(--neutral-light-2);
-  font-size: var(--font-xs);
-  font-weight: var(--font-medium);
-
-  display: flex;
-  align-items: center;
-
-  transition: all ease-in-out 0.04s;
-}
-
-/* stylelint-disable-next-line selector-pseudo-class-no-unknown */
-.createPlaylistButton > *:first {
-  margin-right: 4px;
-}
-
-.createPlaylistButton g path {
-  transition: all ease-in-out 0.04s;
-  fill: var(--neutral-light-2);
-}
-
-.createPlaylistButton:hover {
-  color: var(--neutral);
-  transform: scale(1.02);
-}
-.createPlaylistButton:hover g path {
-  fill: var(--neutral);
-}
-
-.createPlaylistButton:active {
-  transform: scale(1);
-}
