+++ packages/web/src/components/edit-folder-modal/EditFolderModal.tsx
@@ -18,7 +18,7 @@ import { useDispatch } from 'react-redux'
 
 import { useModalState } from 'common/hooks/useModalState'
 import { make, useRecord } from 'common/store/analytics/actions'
-import FolderForm from 'components/create-playlist/FolderForm'
+import FolderForm from 'components/edit-folder-modal/FolderForm'
 import { DeleteFolderConfirmationModal } from 'components/nav/desktop/PlaylistLibrary/DeleteFolderConfirmationModal'
 import { getFolderId } from 'store/application/ui/editFolderModal/selectors'
 import { setFolderId } from 'store/application/ui/editFolderModal/slice'
